+++ scraper/src/youtube2zim/scraper.py
@@ -18,6 +18,7 @@ import tempfile
 from gettext import gettext as _
 from pathlib import Path
 
+import isodate
 import yt_dlp
 import yt_dlp.utils
 from kiwixstorage import KiwixStorage
@@ -1177,8 +1178,11 @@ class Youtube2Zim:
             channel_data = get_channel_json(playlist.creator_id)
             videos = get_videos_list(playlist)
             playlist_videos = [generate_video_preview_object(video) for video in videos]
-
+            playlist_duration = 0
             # add videos to ZIM index
+            for video in videos:
+                video_id = video["contentDetails"]["videoId"]
+                playlist_duration += videos_channels[video_id]["duration_seconds"]
             for idx, video_obj in enumerate(playlist_videos):
                 self.add_custom_item_to_zim_index(
                     video_obj.title,
@@ -1206,10 +1210,17 @@ class Youtube2Zim:
                 thumbnail_path=get_thumbnail_path(
                     videos[0]["contentDetails"]["videoId"]
                 ),
+                duration=isodate.duration_isoformat(
+                    datetime.timedelta(seconds=playlist_duration)
+                ),
             )
 
         def generate_playlist_preview_object(playlist) -> PlaylistPreview:
             videos = get_videos_list(playlist)
+            playlist_duration = 0
+            for video in videos:
+                video_id = video["contentDetails"]["videoId"]
+                playlist_duration += videos_channels[video_id]["duration_seconds"]
             return PlaylistPreview(
                 slug=get_playlist_slug(playlist),
                 id=playlist.playlist_id,
@@ -1219,6 +1230,9 @@ class Youtube2Zim:
                 ),
                 videos_count=len(videos),
                 main_video_slug=get_video_slug(videos[0]),
+                duration=isodate.duration_isoformat(
+                    datetime.timedelta(seconds=playlist_duration)
+                ),
             )
 
         def get_playlist_slug(playlist) -> str: