+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added Total playlist duration in the Playlist view and Playlist panel. (#435)
+
 ## [3.5.0] - 2025-11-17
 
 ### Added